+++ Tests/SwiftLintFrameworkTests/LineEndingTests.swift
@@ -1,0 +1,12 @@
+@testable import SwiftLintFramework
+import XCTest
+
+class LineEndingTests: XCTestCase {
+    func testCarriageReturnDoesNotCauseError() {
+        XCTAssert(
+            violations(
+                "//swiftlint:disable all\r\nprint(123)\r\n"
+            ).isEmpty
+        )
+    }
+}